API_Bookings/requirements.txt
flask
flask-cors
mysqlclient
python-dotenv
flask-jwt-extended
requests